refactor: detailPage 컴포넌트에서 api 호출로 변경
